Update index.html
https://www.accessiplus.com/

"Accessi+ : Our intelligent widget"

"We have developed an accessibility widget for the users of your website or mobile app. Each user can choose a preset profile to enhance the navigation and user experience. The main disabilities encountered on the web have a corresponding preset in our widget. Accessi+ also offers several customization options as well as a screen reader."